Fix lost inode
Filesystem uses blocknr variable that stores a block number that
is a inode itself belongs. When using inodes that are released
in the past, filesystem didn't store anything into blocknr variable.
This caused lost inodes, which filesystem couldn't look up the place on pmem.
Fix this problem to remember the block number when the inode is
released, and then use it when allocating it to a new one.