Fix rmdir method
Not had judged whether a rmdir target directory is empty or not
in the case that dram dentry has private pointer to a direntry on pmem.
This causes removing a directory without removing contents in the
directory. File system tries to remove remaining contents after that but
it has no way of getting them.