Fix spec of aeon_add_dentry()
It had been doing Error handling by judging from whether a returned address was
less than zero or not. The type of variable storing a returned address
is unsigned so it is no meaning. To do error handling correctly, change
spec of aeon_add_dentry() to returning error code and passing a variable
address as an argument that is for storing dentry address.